# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智能墨水屏平台 API
  version: 1.0.0
  description: |
    用户端/管理端/设备端统一接口文档。
    设备（真机与模拟）推荐统一启动链路：login -> auto-register -> bind/status -> login -> stream。
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Firmware:
      type: object
      properties:
        id: { type: string }
        version: { type: string }
        releaseNote: { type: string }
        deviceType: { type: string }
        fileName: { type: string }
        fileUrl: { type: string }
        gridId: { type: string }
        fileSize: { type: integer }
        mime: { type: string }
        sha256: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
...
paths:
...
  /api/firmware:
    get:
      summary: 固件列表
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: ok
    post:
      summary: 新增固件
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Firmware' }
      responses:
        '200':
          description: ok
  /api/firmware/upload:
    post:
      summary: 上传固件文件并创建记录
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                version: { type: string }
                deviceType: { type: string }
                releaseNote: { type: string }
                file: { type: string, format: binary }
      responses:
        '200':
          description: ok
  /api/firmware/upgrade:
    post:
      summary: 单设备升级
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                firmwareId: { type: string }
                force: { type: boolean }
                scheduledAt: { type: string }
      responses:
        '200':
          description: ok
  /api/firmware/upgrades:
    get:
      summary: 升级任务列表
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: deviceId
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        '200':
          description: ok
  /api/firmware/batch-upgrade:
    post:
      summary: 批量升级
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceIds: { type: array, items: { type: string } }
                clusterIds: { type: array, items: { type: string } }
                firmwareId: { type: string }
                force: { type: boolean }
                scheduledAt: { type: string }
      responses:
        '200':
          description: ok
  /api/firmware/upgrades/{jobId}/run:
    post:
      summary: 执行定时升级
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
  /api/firmware/{firmwareId}/delete:
    post:
      summary: 删除固件
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: firmwareId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
  /api/firmware/{firmwareId}/download:
    get:
      summary: 下载固件文件
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: firmwareId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: ok
...